rename thunk actions file
